+++ src/features/like/iLikesContent.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Feed, Segment } from 'semantic-ui-react';
-import { src } from './LikesPanel';
+import { src } from '../like/LikesPanel';
 import { useSelector } from 'react-redux'
-import { selectAllLikes } from './features/like/likeSlice'
+import { selectAllLikes } from './likeSlice'
 
-export default function WeLikesContent() {
+export default function ILikesContent() {
     const likes = useSelector(selectAllLikes)
     return likes.likes.map((item) => <Segment key={item._id}>
         <Feed>
@@ -18,7 +18,7 @@ export default function WeLikesContent() {
                 <Feed.Content>
                     <Feed.Date content={item.name.first + " " + item.name.last + " " + item.age} />
                     <Feed.Summary>
-                        Mы поставили лайк <Feed.User as='p'>{item.name.first}</Feed.User>
+                        Вы поставили лайк <Feed.User as='p'>{item.name.first}</Feed.User>
                         <Feed.Date>{item.registered} </Feed.Date>
                     </Feed.Summary>
                 </Feed.Content>
